refactor(renderer): extract construction path into construction.rs
Move the build/rebuild methods out of spatial_renderer/mod.rs into a new
construction.rs submodule, separating the one-time / rare build path from
the per-frame render hot path:

- new (constructor)
- build_live_params_and_log, build_crossover, build_unified_table
- finish_construction
- refresh_crossover_for_topology (pub(super); the only one called from
  render_frame, once per frame)

These are &mut self / associated methods that do NOT hold the live or
channel-state guards, so they extract cleanly as a second `impl
SpatialRenderer` block (unlike render_frame). fill_band_gains stays in
mod.rs (it is a render-path helper). tests.rs gains explicit imports for
the types it constructs, so mod.rs only imports what its own code uses.

mod.rs: 1814 -> 1255 lines. No behavior change (pure relocation); verified
with the render_frame Criterion bench — deltas are run-to-run noise (a
zero-change re-run moved the same cases by similar amounts). 95/95 renderer
lib tests pass; build + fmt clean.
